test: add unit tests for all microservices
- Gateway: AuthTest (register, login, logout, protected routes)
- Usuarios: 15 tests (seguridad, listar, crear, obtener, actualizar, eliminar)
- Productos: 14 tests (seguridad, listar, crear, obtener, actualizar, eliminar)
- Ordenes: 5 tests (seguridad, crear, listar, validacion, 404)
- Pagos: 10 tests (seguridad, listar, crear, obtener, validacion)
- Notificaciones: 11 tests (seguridad, listar, crear, marcar leida, eliminar)

+++ productos/models.py
@@ -7,7 +7,7 @@ class Producto(db.Model):
     id          = db.Column(db.Integer, primary_key=True)
     nombre      = db.Column(db.String(255), nullable=False)
     descripcion = db.Column(db.Text, default='')
-    precio      = db.Column(db.Numeric(10, 2), nullable=False)
+    precio = db.Column(db.Float, nullable=False)
     stock       = db.Column(db.Integer, default=0)
     categoria   = db.Column(db.String(100), default='')
     activo      = db.Column(db.Boolean, default=True)
